refactor(server): extract PlayerStateManager from ArenaRoom
Extracted player lifecycle management into a new module:
- handlePlayerDeath: death processing, kill tracking, orb spawning
- handlePlayerRespawn: respawn logic, state reset, spawn point
- updatePlayerLevel: level calculation, slot unlocks, talents

Made public: clearBoost, awardKillMass, spawnDeathExplosion,
spawnDeathNeedles, spawnToxicPool, findSpawnPoint, getPlayerRadius,
tryGenerateNextCard, awardTalentToPlayer
